III-7181 migrate Notification component to tailwind
